+++ app/Http/Controllers/ItemCategoryView.php
@@ -41,7 +41,7 @@ class ItemCategoryView extends Controller
 
         $cache_control = new Cache\Control(
             $this->user_id,
-            in_array($resource_type_id, $this->permitted_resource_types, true)
+            in_array((int) $resource_type_id, $this->permitted_resource_types, true)
         );
         $cache_control->setTtlOneWeek();