+++ app/Http/Controllers/Summary/SubcategoryView.php
@@ -39,7 +39,7 @@ class SubcategoryView extends Controller
 
         $cache_control = new Cache\Control(
             $this->user_id,
-            in_array($resource_type_id, $this->permitted_resource_types, true)
+            in_array((int) $resource_type_id, $this->permitted_resource_types, true)
         );
         $cache_control->setTtlOneMonth();