+++ app/Models/Category.php
@@ -58,7 +58,7 @@ class Category extends Model
             join("resource_type", "category.resource_type_id", "resource_type.id")->
             where('category.resource_type_id', '=', $resource_type_id);
 
-        $collection = Clause::applyResourceTypeCollectionCondition(
+        $collection = Clause::applyPermittedResourceTypes(
             $collection,
             $permitted_resource_types,
             $include_public
@@ -114,7 +114,7 @@ class Category extends Model
         join("resource_type", "category.resource_type_id", "resource_type.id")->
         where('category.resource_type_id', '=', $resource_type_id);
 
-        $collection = Clause::applyResourceTypeCollectionCondition(
+        $collection = Clause::applyPermittedResourceTypes(
             $collection,
             $permitted_resource_types,
             $include_public