+++ app/Models/ResourceType.php
@@ -81,7 +81,7 @@ class ResourceType extends Model
     {
         $collection = $this->select("resource_type.id");
 
-        $collection = Clause::applyResourceTypeCollectionCondition(
+        $collection = Clause::applyPermittedResourceTypes(
             $collection,
             $permitted_resource_types,
             $include_public
@@ -127,6 +127,7 @@ class ResourceType extends Model
                 'resource_type.public AS resource_type_public',
                 'item_type.id AS resource_type_item_type_id',
                 'item_type.name AS resource_type_item_type_name',
+                'item_type.friendly_name AS resource_type_item_type_friendly_name',
                 'item_type.description AS resource_type_item_type_description'
             )
             ->selectRaw('
@@ -143,7 +144,7 @@ class ResourceType extends Model
             ->join('item_type', 'resource_type_item_type.item_type_id', 'item_type.id')
             ->leftJoin("resource", "resource_type.id", "resource.id");
 
-        $collection = Clause::applyResourceTypeCollectionCondition(
+        $collection = Clause::applyPermittedResourceTypes(
             $collection,
             $permitted_resource_types,
             $include_public
@@ -197,6 +198,7 @@ class ResourceType extends Model
                 'resource_type.public AS resource_type_public',
                 'item_type.id AS resource_type_item_type_id',
                 'item_type.name AS resource_type_item_type_name',
+                'item_type.friendly_name AS resource_type_item_type_friendly_name',
                 'item_type.description AS resource_type_item_type_description'
             )
             ->selectRaw('
@@ -243,7 +245,7 @@ class ResourceType extends Model
                 'resource_type.description AS resource_type_description'
             );
 
-        $collection = Clause::applyResourceTypeCollectionCondition(
+        $collection = Clause::applyPermittedResourceTypes(
             $collection,
             $permitted_resource_types,
             $include_public
@@ -270,6 +272,7 @@ class ResourceType extends Model
             'resource_type_public' => $resource_type->public,
             'resource_type_item_type_id' => $resource_type->item_type->id,
             'resource_type_item_type_name' => $resource_type->item_type->name,
+            'resource_type_item_type_friendly_name' => $resource_type->item_type->friendly_name,
             'resource_type_item_type_description' => $resource_type->item_type->description
         ];
     }